chore: move files to be consistent with other routes structures
