+++ apps/token/src/routes/staking/disassociate/components/disassociate-page/index.ts
@@ -1,0 +1,1 @@
+export { DisassociatePage } from './disassociate-page';
