+++ apps/token/src/routes/staking/disassociate/components/disassociate-transaction/index.tsx
@@ -1,0 +1,1 @@
+export { DisassociateTransaction } from './disassociate-transaction';
