+++ apps/token/src/routes/staking/disassociate/disassociate-page-container.tsx
@@ -1,7 +1,7 @@
 import { useVegaWallet } from '@vegaprotocol/wallet';
 import { useWeb3React } from '@web3-react/core';
 import { EthConnectPrompt } from '../../../components/eth-connect-prompt';
-import { DisassociatePage } from './disassociate-page';
+import { DisassociatePage } from './components/disassociate-page';
 
 export const DisassociateContainer = () => {
   const { account } = useWeb3React();
